pyproject.toml
[project]
name = "reposentinel"
version = "1.0.0"
description = "Repo intelligence engine for Copilot guardrails"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "tree-sitter>=0.20.0",
    "rich>=13.0.0",
    "tree-sitter-languages>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
reposentinel = "cmd.reposentinel.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
